float to double

+++ cocos/core/math/math-base.ts
@@ -1,7 +1,7 @@
 import { ValueType } from '../value-types/value-type';
 import { FloatArray } from './type-define';
 
-export const MATH_FLOAT_ARRAY = Float32Array;
+export const MATH_FLOAT_ARRAY = Float64Array;
 
 export class MathBase extends ValueType {
     public static createFloatArray (size: number) {
